fix: Get ticket support type for forwarding
When an individual asked for psycologic and legal support, the tickets
were filtered by 'tipo_de_acolhimento' from the user data. If the user
asks for two types of support, two tickets are created, but if it was
only filtered by the 'tipo_de_acolhimento', it couldnt know what type of
support the ticket was created for. This is fixed by getting the type of
support from the ticket subject, if the user field 'tipo_de_acolhimento' asks
for the two types.